Minor tweaks to DiscordProvider
Changed how Connected is determined. If we're in the process of connecting or disconnecting, let the library finish those operations before attempting our reconnect.

The endless jobs are a lack of client.Ready never triggering. Adjusted the call order to address what looks like a potential deadlock scenario for this.

Hopefully fixes #1250