Discord replies no longer mention the user.
I don't want this clogging up my mentions inbox (which i check for things i forgot to do) every time i do !tgs. and it will once there are 7 bots on tgs 5 and all 7 of them mention me every time I use a command.

There is no need to mention or ping the user for a response to a message they sent less than a second ago. 

The goal of a mention is to alert the user there is a message they might want to see, but in this case, they kinda know, given they just sent the command.